\name{ACtest-methods}
\alias{print.ACtest}

\title{
Methods for class ACtest
}
\description{
\code{print.ACtest()} prints the P. values of the test.
}
\usage{
\method{print}{ACtest}(x, \dots)
}

\arguments{
  \item{x}{
an object of class 'ACtest' (as returned by the function \code{\link{ACtest}}).
}
  \item{\dots}{
  additional arguments to \code{\link{print}}.
}
}