\name{DataFiles}

\alias{DataFiles}

\alias{VodafoneCDS}

\title{Multiple Time Series Data Set} 

\description{
    The data matrix \code{VodafoneCDS} is used for the examples. It consists of 804 daily observations, from 1 January 2009 to 31 January 2012, of Vodafone's Credit Default Swap prices (the 'CDS' column) and the credit spread on its bond over the risk-free rate (the 'SWSP' column). For more information, see Ahlgren and Catani (2016), who used the same data set.
	}
	
\references{
Ahlgren, N. & Catani, P. (2016).
\emph{Wild bootstrap tests for autocorrelation in vector autoregressive models.} Stat Papers, <doi:10.1007/s00362-016-0744-0>.
}