\name{wildBoot-methods}
\alias{print.wildBoot}

\title{
Methods for class wildBoot
}
\description{
\code{print.wildBoot()} prints the P. values of the WB test.
}
\usage{
\method{print}{wildBoot}(x, \dots)
}

\arguments{
  \item{x}{
an object of class 'wildBoot' (as returned by the function \code{\link{wildBoot}}).
}
  \item{\dots}{
  additional arguments to \code{\link{print}}.
}
}